requirements.txt
Flask==3.0.2
Flask-PyMongo==2.3.0
pymongo==4.5.0
qrcode==7.3
pillow==10.0.0
python-dotenv==1.0.0
werkzeug
gunicorn==20.1.0
itsdangerous
Jinja2
